+++ examples/raspberry_pi.rs
@@ -1,10 +1,10 @@
 use {
     display_interface_spi::SPIInterface,
     embedded_graphics::{
-        fonts::{Font24x32, Text},
+        mono_font::{ascii::FONT_10X20, MonoTextStyleBuilder},
         pixelcolor::Gray4,
         prelude::*,
-        text_style,
+        text::{Baseline, Text},
     },
     linux_embedded_hal::Delay,
     rppal::{
@@ -34,13 +34,13 @@ fn main() {
     disp.clear(Gray4::new(0)).unwrap();
     disp.flush().unwrap();
 
+    let text_style = MonoTextStyleBuilder::new()
+        .font(&FONT_10X20)
+        .text_color(Gray4::new(0b0000_1111))
+        .build();
+
     // Write "Hello" to the display
-    Text::new("Hello", Point::new(0, 0))
-        .into_styled(text_style!(
-            font = Font24x32,
-            text_color = Gray4::new(0b0000_1111),
-            background_color = Gray4::new(0),
-        ))
+    Text::with_baseline("Hello", Point::zero(), text_style, Baseline::Top)
         .draw(&mut disp)
         .unwrap();
     disp.flush().unwrap();