+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## v0.2.0 - In Progress
 
  - Fix issue where pixels luma is never set back to 0
+ - Migrate to embedded-graphics-core 0.3 (embedded-graphics-0.7 API).
+ - Migrate examples to embedded-graphics 0.7 API.
+ - Add bounding box check
 
 ## v0.1.0 - 2020-11-25